requirements.txt
Django==3.1.13
django-crispy-forms==1.13.0
httpx==0.21.1
mysqlclient==2.1.0
python-decouple==3.5
